requirements.txt
Flask
requests
streamlit
python-dotenv

